move graphcontroller to controllers
