+++ packages/ReactDagEditor/src/controllers/GraphController.ts
@@ -3,7 +3,7 @@ import { GraphFeatures } from "../Features";
 import { IPoint } from "../models/geometry";
 import { NodeModel } from "../models/NodeModel";
 import { GraphBehavior, IGraphState } from "../models/state";
-import { EventChannel } from "./eventChannel";
+import { EventChannel } from "../utils/eventChannel";
 
 /**
  * event handlers must get/set GraphBehavior immediately
