+++ packages/ReactDagEditor/src/handlers/container.ts
@@ -7,7 +7,7 @@ import { GraphCanvasEvent, ICanvasCommonEvent, IEvent } from "../models/event";
 import { IPoint } from "../models/geometry";
 import { CanvasMouseMode, IGraphState } from "../models/state";
 import { EventChannel } from "../utils/eventChannel";
-import { GraphController } from "../utils/graphController";
+import { GraphController } from "../controllers/GraphController";
 import { isWithinThreshold } from "../utils/isWithinThreshold";
 
 export interface IContainerMouseDownParams {
